requirements.txt
alpaca-trade-api>=2.3.0
pandas>=1.3.0
numpy>=1.20.0
python-dotenv>=0.19.0
schedule>=1.1.0
ta>=0.10.1
matplotlib>=3.5.0
requests>=2.25.0
yfinance>=0.2.4  # Added as a fallback data source
lightweight-charts>=0.1.0  # For interactive charting